refactor: fe seperate queryclient to module directory
